+++ CHANGELOG.md
@@ -2,7 +2,12 @@
 # Sunder counter and reminder
 # Battle Shout reminder
 
-## v1.0.1 (15.01.2025)
+## v1.0.2 (18.01.2025)
+changed debuff checking to check all debuffs
+only show ap & sunder if target is hostile
+Change update from eventbased to a timer
+
+## v1.0.1 (17.01.2025)
 Added missing code from initial commit
 Added slashcommands to hide/show the individual frames